requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
colorama==0.4.4
decouple==0.0.7
discord==1.7.3
discord.py==1.7.3
future==0.18.2
idna==3.3
multidict==5.2.0
pefile==2021.9.3
Pillow==9.0.0
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
PyNaCl==1.5.0
pywin32-ctypes==0.2.0
requests==2.27.1
rule34==1.8.1
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
youtube-dl==2021.12.17
